split gossip conformance into agave-conformance crate
  Move gossip_decode_to_effects and the sol_compat_gossip_decode_v1 FFI
  entry point out of solana-gossip into a new agave-conformance crate in
  the dev-bins workspace. Expose the internals the harness needs under
  dev-context-only-utils via qualifier_attr, and drop the now-unused
  conformance feature (and prost/protosol deps) from solana-gossip.
