+++ dev-bins/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["../accounts-db/store-tool", "../ledger-tool"]
+members = ["../accounts-db/store-tool", "../conformance", "../ledger-tool"]
 
 resolver = "2"
 
@@ -41,6 +41,7 @@ agave-snapshots = { path = "../snapshots", version = "=4.2.0-alpha.0", features 
 ahash = "0.8.11"
 assert_cmd = "2.2.2"
 bincode = "1.3.3"
+bv = "0.11.1"
 chrono = { version = "0.4.42", default-features = false }
 clap = { version = "2.33.1", default-features = false, features = ["suggestions"] }
 crossbeam-channel = "0.5.15"
@@ -58,7 +59,7 @@ num_cpus = "1.17.0"
 pretty-hex = "0.4.2"
 prost = "0.14.4"
 prost-build = "0.14.4"
-protosol = "2.0.0"
+protosol = "=8.2.0"
 rand = "0.9.4"
 rayon = "1.12.0"
 regex = "1.12.4"
@@ -71,6 +72,7 @@ signal-hook = "0.4.4"
 solana-account = "4.3.1"
 solana-account-decoder = { path = "../account-decoder", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-accounts-db = { path = "../accounts-db", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
+solana-bloom = { path = "../bloom", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-bpf-loader-program = { path = "../programs/bpf_loader", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-builtins = { path = "../builtins", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-clap-utils = { path = "../clap-utils", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
@@ -94,6 +96,7 @@ solana-genesis = { path = "../genesis", version = "=4.2.0-alpha.0", features = [
 solana-genesis-config = "4.0.0"
 solana-genesis-utils = { path = "../genesis-utils", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-geyser-plugin-manager = { path = "../geyser-plugin-manager", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
+solana-gossip = { path = "../gossip", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-hash = "4.4.0"
 solana-inflation = "3.1.1"
 solana-instruction = "3.4.0"
@@ -109,6 +112,7 @@ solana-metrics = { path = "../metrics", version = "=4.2.0-alpha.0", features = [
 solana-native-token = "3.0.0"
 solana-net-utils = { path = "../net-utils", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-nonce = "3.2.0"
+solana-packet = "4.1.0"
 solana-precompile-error = "3.0.0"
 solana-program-runtime = { path = "../program-runtime", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-pubkey = { version = "4.2.0", default-features = false }
@@ -119,6 +123,7 @@ solana-rpc-client = { path = "../rpc-client", version = "=4.2.0-alpha.0", defaul
 solana-rpc-client-api = { path = "../rpc-client-api", version = "=4.2.0-alpha.0" }
 solana-runtime = { path = "../runtime", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
 solana-runtime-transaction = { path = "../runtime-transaction", version = "=4.2.0-alpha.0", features = ["agave-unstable-api"] }
+solana-sanitize = "3.0.1"
 solana-sbpf = { version = "=0.21.1", default-features = false }
 solana-sdk-ids = "3.1.0"
 solana-shred-version = "3.0.1"
@@ -150,6 +155,7 @@ solana-vote-program = { path = "../programs/vote", version = "=4.2.0-alpha.0", d
 tempfile = "3.23.0"
 thiserror = "2.0.17"
 tokio = "1.48.0"
+wincode = { version = "0.5.5", features = ["derive"] }
 
 [profile.dev]
 debug = "line-tables-only"
