+++ gossip/Cargo.toml
@@ -33,8 +33,7 @@ frozen-abi = [
 ]
 agave-unstable-api = []
 dev-context-only-utils = ["solana-bloom/dev-context-only-utils"]
-conformance = ["dep:prost", "dep:protosol", "dev-context-only-utils"]
-dummy-for-ci-check = ["conformance"]
+dummy-for-ci-check = ["dev-context-only-utils"]
 
 [dependencies]
 agave-logger = { workspace = true }
@@ -50,8 +49,7 @@ itertools = { workspace = true }
 lazy-lru = { workspace = true }
 log = { workspace = true }
 num-traits = { workspace = true }
-prost = { version = "0.11", optional = true }
-protosol = { workspace = true, optional = true }
+qualifier_attr = { workspace = true }
 rand = { workspace = true }
 rand_chacha = { workspace = true }
 rayon = { workspace = true }
@@ -104,7 +102,7 @@ bs58 = { workspace = true }
 criterion = { workspace = true }
 num_cpus = { workspace = true }
 serial_test = { workspace = true }
-solana-gossip = { path = ".", features = ["agave-unstable-api", "dev-context-only-utils", "conformance"] }
+solana-gossip = { path = ".", features = ["agave-unstable-api", "dev-context-only-utils"] }
 solana-net-utils = { path = "../net-utils", features = ["agave-unstable-api", "dev-context-only-utils"] }
 solana-perf = { path = "../perf", features = ["agave-unstable-api", "dev-context-only-utils"] }
 solana-runtime = { path = "../runtime", features = ["agave-unstable-api", "dev-context-only-utils"] }
