+++ gossip/src/duplicate_shred.rs
@@ -18,6 +18,8 @@ use {
     thiserror::Error,
     wincode::{ReadError, SchemaRead, SchemaWrite, WriteError},
 };
+#[cfg(feature = "dev-context-only-utils")]
+use qualifier_attr::qualifiers;
 
 const DUPLICATE_SHRED_HEADER_SIZE: usize = 63;
 
@@ -52,35 +54,42 @@ pub struct DuplicateShred {
 
 impl DuplicateShred {
     #[inline]
+    #[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
     pub(crate) fn num_chunks(&self) -> u8 {
         self.num_chunks
     }
 
     #[inline]
+    #[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
     pub(crate) fn chunk_index(&self) -> u8 {
         self.chunk_index
     }
 
-    // Conformance-only accessors; unused under DCOU.
-    #[cfg(any(test, feature = "conformance"))]
+    // Accessors exposed under `dev-context-only-utils` for the out-of-crate
+    // conformance harness (see the `agave-conformance` crate).
+    #[cfg(any(test, feature = "dev-context-only-utils"))]
+    #[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
     #[inline]
     pub(crate) fn from(&self) -> &Pubkey {
         &self.from
     }
 
-    #[cfg(any(test, feature = "conformance"))]
+    #[cfg(any(test, feature = "dev-context-only-utils"))]
+    #[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
     #[inline]
     pub(crate) fn wallclock(&self) -> u64 {
         self.wallclock
     }
 
-    #[cfg(any(test, feature = "conformance"))]
+    #[cfg(any(test, feature = "dev-context-only-utils"))]
+    #[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
     #[inline]
     pub(crate) fn slot(&self) -> Slot {
         self.slot
     }
 
-    #[cfg(any(test, feature = "conformance"))]
+    #[cfg(any(test, feature = "dev-context-only-utils"))]
+    #[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
     #[inline]
     pub(crate) fn chunk(&self) -> &[u8] {
         &self.chunk
